fix nonetype metadata
Issue: AAH-2296

+++ galaxy_ng/app/models/namespace.py
@@ -75,11 +75,12 @@ class Namespace(
     @avatar_url.setter
     def avatar_url(self, value):
 
-        pk = self.last_created_pulp_metadata.pk
-        # allow to rebuild latest namespace metadata
-        if metadata := AnsibleNamespaceMetadata.objects.filter(pk=pk).first():
-            metadata.delete()
-        self.last_created_pulp_metadata = None
+        if self.last_created_pulp_metadata:
+            pk = self.last_created_pulp_metadata.pk
+            # allow to rebuild latest namespace metadata
+            if metadata := AnsibleNamespaceMetadata.objects.filter(pk=pk).first():
+                metadata.delete()
+            self.last_created_pulp_metadata = None
 
         self._avatar_url = value
